requirements.txt
numpy
pandas
scikit-learn
tensorflow>=2.12
opencv-python
Pillow
